backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
werkzeug==2.3.7
pillow==10.0.0
bcrypt==4.0.1
pydantic==2.0.3
python-multipart==0.0.6 